aaaaaahhhhhhhhh authorization + fixes

+++ server/src/routes/express/allowed_hosts.ts
@@ -1,11 +1,37 @@
 import express, { Request } from "express";
-import { AllowedHost, Database } from "../../databases/base.js";
+import { Database } from "../../databases/base.js";
 import logger from "../../logger/bunyan.js";
+import { AllowedHost, PolicyResult } from "../../schema/index.js";
+import { Authorization } from "../../authorization/base.js";
 
-export function buildAllowedHostsRoutes(db: Database) {
+export function buildAllowedHostsRoutes(db: Database, authorizer?: Authorization) {
   const router = express.Router();
+  router.use(async (req, res, next) => {
+    if(authorizer){
+      const { body, params } = req
+      const { user_id: userId } = body
+      if(userId){
+        const policies = await authorizer.getAllowedHostsPolicies(userId, { body, params })
+        body.policies = policies;
+        next()
+      }else{
+        logger.error("Failed allowed hosts request: Unauthorized empty user")
+        res.status(403).send({ error: "Failed allowed hosts request: Unauthorized empty user" })
+        return
+      }
+    }else{
+      next()
+    }
+  })
+  
   router.delete("/:id", async (req, res) => {
     try {
+      const { body: { policies } } = req
+      if(!policies.deleteAllowed){
+        logger.error("Failed to delete allowed hosts: Unauthorized!");
+        res.status(500).send(JSON.stringify({ error: "Failed to delete allowed hosts: Unauthorized!" }));
+        return;
+      }
       const { id } = req.params;
       await db.deleteAllowedHost(id);
       logger.info(`Deleted allowed hosts: ${id}`);
@@ -17,10 +43,38 @@ export function buildAllowedHostsRoutes(db: Database) {
       return;
     }
   }); // delete allowed hosts
-
-  router.get("/", async (_, res) => {
+  
+  router.post("/:id", async (req, res) => {
     try {
-      const allowedHosts = await db.getAllowedHosts();
+      const { body: { policies } } = req
+      if(!policies.updateAllowed){
+        logger.error("Failed to update allowed hosts: Unauthorized!");
+        res.status(500).send(JSON.stringify({ error: "Failed to update allowed hosts: Unauthorized!" }));
+        return;
+      }
+      const { body, params } = req;
+      await db.updateAllowedHost(params.id, body.hosts);
+      logger.info(`Set allowed hosts to: ${JSON.stringify(body.hosts)}`);
+      res.status(200).send({ status: "ok" });
+      return;
+    } catch (error) {
+      logger.error("Failed to update allowed hosts:", error);
+      res.status(500).send(JSON.stringify(error));
+      return;
+    }
+  })
+
+  router.get("/", async (req, res) => {
+    try {
+      const { body } = req;
+      const { user_id: userId } = body
+      const { body: { policies } } = req
+      if(!policies.readAllAllowed && !userId){
+        logger.error("Failed to fetch allowed hosts: Unauthorized!");
+        res.status(500).send(JSON.stringify({ error: "Failed to fetch allowed hosts: Unauthorized!" }));
+        return;
+      }
+      const allowedHosts = await db.getAllowedHosts(userId);
       logger.info(`Fetched all allowed hosts: ${JSON.stringify(allowedHosts)}`);
       res.status(200).send({ allowed_hosts: allowedHosts });
       return;
@@ -33,10 +87,16 @@ export function buildAllowedHostsRoutes(db: Database) {
 
   router.post(
     "/",
-    async (req: Request<{}, {}, { hosts: AllowedHost[] }>, res) => {
+    async (req: Request<{}, {}, { hosts: Omit<AllowedHost, "id">[], policies: PolicyResult }>, res) => {
       try {
+        const { body: { policies } } = req
+        if(!policies.createAllowed){
+          logger.error("Failed to set allowed hosts: Unauthorized!");
+          res.status(500).send(JSON.stringify({ error: "Failed to set allowed hosts: Unauthorized!" }));
+          return;
+        }
         const { body } = req;
-        await db.setAllowedHosts(body.hosts);
+        await db.createAllowedHosts(body.hosts);
         logger.info(`Set allowed hosts to: ${JSON.stringify(body.hosts)}`);
         res.status(200).send({ status: "ok" });
         return;
